Cleanup and reorg of nested_multi
