+++ CMakeLists.txt
@@ -74,6 +74,7 @@ set (raja_sources
   src/DepGraphNode.cpp
   src/LockFreeIndexSetBuilders.cpp
   src/MemUtils_CUDA.cpp
+	src/SharedMemory.cpp
   src/ThreadUtils_CPU.cpp)
 
 set (raja_depends)
