+++ include/RAJA/policy/cuda/nested.hpp
@@ -9,11 +9,14 @@
  ******************************************************************************
  */
 
+#include "RAJA/pattern/nested.hpp"
+#include "RAJA/pattern/nested_multi.hpp"
+
 #ifndef RAJA_policy_cuda_nested_HPP
 #define RAJA_policy_cuda_nested_HPP
 
 #include "RAJA/config.hpp"
-#include "RAJA/pattern/nested.hpp"
+
 
 #if defined(RAJA_ENABLE_CUDA)
 
@@ -62,10 +65,11 @@
 #include <cassert>
 #include <climits>
 
-#include "RAJA/RAJA.hpp"
 #include "RAJA/config.hpp"
 #include "RAJA/util/defines.hpp"
 #include "RAJA/util/types.hpp"
+
+
 
 #include "RAJA/policy/cuda/MemUtils_CUDA.hpp"
 #include "RAJA/policy/cuda/policy.hpp"
@@ -342,42 +346,6 @@ struct Executor<Collapse<cuda_collapse_exec<Async>, FOR_TYPES...>> {
     return dims;
   }
 
-//  template <typename WrappedBody, typename BeginTuple>
-//  RAJA_INLINE
-//  auto createDeviceWrapper(
-//      CudaDim &dims,
-//      BeginTuple const &begin_tuple,
-//      WrappedBody const &wrap) ->
-//
-//      ForWrapper<WrappedBody,
-//                  BeginTuple,
-//                 typename FOR_TYPES::policy_type::cuda_exec_policy...>
-//
-//  {
-//
-//    /* As we create a cuda wrapper, we construct all of the cuda loop policies,
-//     * like cuda_thread_x_exec, their associated segment from wrap.data.st
-//     *
-//     * This construction modifies the CudaDim to specify the correct number
-//     * of threads and blocks for the kernel launch
-//     *
-//     * The wrapped body is the device function to be launched, and does all
-//     * of the block/thread idx unpacking and assignment
-//    */
-//    return
-//        ForWrapper<WrappedBody,
-//                   BeginTuple,
-//                   typename FOR_TYPES::policy_type::cuda_exec_policy...>(
-//            wrap,
-//
-//            begin_tuple,
-//
-//            typename FOR_TYPES::policy_type::cuda_exec_policy(
-//                dims, camp::get<FOR_TYPES::index_val>(wrap.data.st))...
-//
-//            );
-//
-//  }
 
 
   template<typename WrappedBody>
@@ -441,7 +409,7 @@ struct Executor<Collapse<cuda_collapse_exec<Async>, FOR_TYPES...>> {
       cudaStream_t stream = 0;
 
       // Get amount of dynamic shared memory requested by SharedMemory objects
-      size_t shmem = RAJA::cuda::detail::shared_memory_total_bytes;
+      size_t shmem = RAJA::detail::getSharedMemorySize();
 
       internal::cudaLauncher<<<dims.num_blocks, dims.num_threads,
           shmem, stream>>>(
@@ -456,183 +424,6 @@ struct Executor<Collapse<cuda_collapse_exec<Async>, FOR_TYPES...>> {
 };
 
 
-
-
-template<bool Async, size_t i, size_t N>
-struct InvokeLoopsCUDA {
-
-  template<typename ... LoopDims, typename ... LoopList>
-  RAJA_DEVICE
-  void operator()(camp::tuple<LoopDims...> const &loop_dims,
-                  camp::tuple<LoopList...> &loops) {
-
-    if(camp::get<i>(loop_dims).threadIncluded()){
-      //invokeLoopData(camp::get<i>(loops));
-      camp::get<i>(loops)();
-    }
-
-    if(!Async){
-      __syncthreads();
-    }
-
-    InvokeLoopsCUDA<Async, i+1, N> next_invoke;
-    next_invoke(loop_dims, loops);
-  }
-
-};
-
-
-template<bool Async, size_t N>
-struct InvokeLoopsCUDA<Async, N, N> {
-
-  template<typename ... LoopDims, typename ... LoopList>
-  RAJA_DEVICE
-  void operator()(camp::tuple<LoopDims...> const &,
-      camp::tuple<LoopList...> &) {}
-
-};
-
-
-
-
-template<bool Async = false>
-struct cuda_multi_exec{};
-
-
-template<typename NestedPolicy, typename SegmentTuple, typename Body>
-auto createLoopExecutor(LoopData<NestedPolicy, SegmentTuple, Body> const &loop_data) ->
-Executor<camp::at_v<NestedPolicy, 0>>
-{
-  // Extract the first policy from the RAJA::nested::Policy
-  // We are assuming that this policy is going to be a CudaCollapse
-  using collapse_policy = camp::at_v<NestedPolicy, 0>;
-
-  // Use the Executor class to compute what thread/block dimensions
-  // are needed for this kernel
-  Executor<collapse_policy> exec;
-
-  return exec;
-}
-
-
-template<typename NestedPolicy, typename SegmentTuple, typename Body>
-CudaDim computeCudaDims(CudaDim &launch_dims,
-    LoopData<NestedPolicy, SegmentTuple, Body> const &loop_data)
-{
-  // Extract the first policy from the RAJA::nested::Policy
-  // We are assuming that this policy is going to be a CudaCollapse
-  using collapse_policy = camp::at_v<NestedPolicy, 0>;
-
-  // Use the Executor class to compute what thread/block dimensions
-  // are needed for this kernel
-  Executor<collapse_policy> exec;
-  CudaDim dims = exec.computeCudaDim(loop_data.st);
-
-  printf("Loop Dims: \n");
-  dims.print();
-
-  // keep track of maximum launch dimensions
-  launch_dims = launch_dims.maximum(dims);
-
-  // Return this kernels launch dims
-  return dims;
-}
-
-template<bool Async, typename LoopDimList, typename LoopList>
-struct CudaMultiWrapper {};
-
-template<bool Async, typename ... LoopDims, typename ... Loops>
-struct CudaMultiWrapper<Async, camp::tuple<LoopDims...>, camp::tuple<Loops...>>{
-  camp::tuple<LoopDims...> loop_dims;
-  camp::tuple<Loops...> loops;
-
-  RAJA_DEVICE void operator()()
-  {
-    InvokeLoopsCUDA<Async, 0, sizeof...(Loops)> invoker;
-    invoker(loop_dims, loops);
-  }
-};
-
-
-
-
-template <bool Async, camp::idx_t... LoopIdx, typename ... LoopList>
-RAJA_INLINE void forall_multi_idx(
-    cuda_multi_exec<Async>,
-    camp::idx_seq<LoopIdx...> const &,
-    LoopList & ... loop_datas)
-{
-
-  // Create a tuple of Executor objects
-
-  auto executors = camp::make_tuple(createLoopExecutor(loop_datas)...);
-
-
-
-  auto loop_tuple = camp::make_tuple(make_cuda_wrapper(loop_datas)...);
-
-
-
-  // Create a tuple of device wrappers from the executors
-  // also, compute shared memory requirements
-  RAJA::detail::startSharedMemorySetup();
-
-  CudaDim foo_dim;
-  auto loop_wraps = camp::make_tuple(
-      camp::get<LoopIdx>(executors).createDeviceWrapper(
-          foo_dim, camp::get<LoopIdx>(loop_tuple)
-      )
-      ...
-  );
-
-  RAJA::detail::finishSharedMemorySetup();
-
-
-
-  // Compute dim3's for thread and blocks, for each loop
-  // launch_dims becomes the max over all of the dimensions
-  CudaDim dims;
-  auto loop_dims = camp::make_tuple(computeCudaDims(dims, loop_datas)...);
-
-
-  printf("Launch Dims: \n");
-  dims.print();
-
-  // Step 3: Wrap the loops with a device-side invoker
-
-  using wrapper_type = CudaMultiWrapper<Async, decltype(loop_dims), decltype(loop_wraps)>;
-  wrapper_type wrap {loop_dims, loop_wraps};
-
-
-  // Step 4: launch our kernel!
-  cudaStream_t stream = 0;
-
-  // Get amount of dynamic shared memory requested by SharedMemory objects
-  size_t shmem = RAJA::cuda::detail::shared_memory_total_bytes;
-  printf("Dynamic shared memory: %ld bytes\n", (long)shmem);
-  shmem = 4096;  // TODO: FIXME!!!
-
-
-  internal::cudaLauncher<<<dims.num_blocks, dims.num_threads,
-      shmem, stream>>>(
-      RAJA::cuda::make_launch_body(
-          dims.num_blocks, dims.num_threads, shmem, stream, wrap));
-  RAJA::cuda::peekAtLastError();
-
-  RAJA::cuda::launch(stream);
-}
-
-template <bool Async, typename ... LoopList>
-RAJA_INLINE void forall_multi(
-    cuda_multi_exec<Async> const &exec,
-    LoopList ... loop_datas)
-{
-
-  using loop_idx = typename camp::make_idx_seq<sizeof...(LoopList)>::type;
-
-    forall_multi_idx(exec, loop_idx{},  loop_datas...);
-
-}
 
 
 }  // namespace nested
