+++ include/RAJA/policy/sequential.hpp
@@ -30,8 +30,10 @@
 
 #include "RAJA/policy/sequential/atomic.hpp"
 #include "RAJA/policy/sequential/forall.hpp"
+#include "RAJA/policy/sequential/nested_multi.hpp"
 #include "RAJA/policy/sequential/policy.hpp"
 #include "RAJA/policy/sequential/reduce.hpp"
 #include "RAJA/policy/sequential/scan.hpp"
+
 
 #endif  // closing endif for header file include guard
