+++ test/unit/CMakeLists.txt
@@ -51,6 +51,10 @@ raja_add_test(
   SOURCES nested.cpp)
 
 raja_add_test(
+  NAME test-nested_multi
+  SOURCES test-nested_multi.cpp)
+
+raja_add_test(
   NAME test-multipolicy
   SOURCES test-multipolicy.cpp)
